Work with teachers added
Authorization for teachers
Access teachers to some pages
Pages versions for students or teachers
Other small updates

+++ StudentRating/src/main/java/com/example/studentrating/controllers/StudentsRestController.java
@@ -36,69 +36,31 @@ public class StudentsRestController {
     @Autowired
     private PastYearPointRepository pastYearPointRepository;
 
-    @PostMapping("/signIn")
-    public String signIn(String login, String password, HttpSession session) {
-        Student student = studentRepository.findByLoginAndPassword(login, encryptText(password));
-        if (student != null) {
-            Session.setSession(session, student.getId(), student.getRole());
-            return "success";
-        } else {
-            return "error";
-        }
-    }
-
-    @PostMapping("/signOut")
-    public String signOut(HttpSession request) {
-        try {
-            Session.signOut(request);
-            return "success";
-        } catch (Exception e) {
-            return "e.getMessage()";
-        }
-    }
-
     @PostMapping("/addNew")
     public String addNew(String surname, String name, String patronymic, String login, String group, @RequestParam(defaultValue = "false") boolean isStudentCouncil, HttpSession session) {
         try {
-            if (studentRepository.findByLogin(login) == null && teacherRepository.findByLogin(login) == null) {
-                Student student = new Student();
-                student.setSurname(surname);
-                student.setName(name);
-                student.setPatronymic(patronymic);
-                student.setLogin(login);
-                student.setGroupName(group);
-                student.setStudentCouncil(isStudentCouncil);
-                student.setPassword(encryptText("123"));
-                student.setImageUrl("/img/avatars/personal.svg");
-                student.setRole("STUDENT");
-                student.setPoints(0);
-                studentRepository.save(student);
-                return "success";
-            } else {
-                return "Введите другой логин";
-            }
-
-//            return student;
-        } catch (Exception e) {
-            return e.getMessage();
-        }
-    }
-
-    @PostMapping("/updateProfile")
-    public String updateProfile(String login, String password1, HttpSession session) {
-        try {
-            // image - null, password1 = ""
-            if (!login.equals("") || password1.equals("")) {
-                Student student = studentRepository.findById(Session.getSessionId(session)).get();
-                if (!login.equals(studentRepository.findById(Session.getSessionId(session)).get().getLogin())) {
-                    if (login != "" && studentRepository.findByLogin(login) == null && teacherRepository.findByLogin(login) == null)
-                        student.setLogin(login);
-                    else return "Введите другой логин";
+            if (Session.isAuthorize(session).equals("ADMIN")) {
+                if (studentRepository.findByLogin(login) == null && teacherRepository.findByLogin(login) == null) {
+                    Student student = new Student();
+                    student.setSurname(surname);
+                    student.setName(name);
+                    student.setPatronymic(patronymic);
+                    student.setLogin(login);
+                    student.setGroupName(group);
+                    student.setStudentCouncil(isStudentCouncil);
+                    student.setPassword(encryptText("123"));
+                    student.setImageUrl("/img/avatars/personal.svg");
+                    student.setRole("STUDENT");
+                    student.setPoints(0);
+                    studentRepository.save(student);
+                    return "success";
+                } else {
+                    return "Введите другой логин";
                 }
-                if (!password1.equals("")) student.setPassword(encryptText(password1));
-                studentRepository.save(student);
+            } else {
+                return "У вас нет прав на данное действие";
             }
-            return "success";
+//            return student;
         } catch (Exception e) {
             return e.getMessage();
         }
@@ -107,35 +69,27 @@ public class StudentsRestController {
     @PostMapping("/deleteById")
     public String deleteById(Long studentId, HttpSession session) {
         try {
-            ArrayList<PastYearPoint> pastYearPoints = pastYearPointRepository.findAllByStudent_Id(studentId);
-            pastYearPointRepository.deleteAll(pastYearPoints);
+            if (Session.isAuthorize(session).equals("ADMIN")) {
+                ArrayList<PastYearPoint> pastYearPoints = pastYearPointRepository.findAllByStudent_Id(studentId);
+                pastYearPointRepository.deleteAll(pastYearPoints);
 
-            ArrayList<Activity> activities = activityRepository.findAllByStudent_Id(studentId);
-            activityRepository.deleteAll(activities);
+                ArrayList<Activity> activities = activityRepository.findAllByStudent_Id(studentId);
+                activityRepository.deleteAll(activities);
 
-            ArrayList<Notification> notifications = notificationRepository.findAllByStudent_Id(studentId);
-            notificationRepository.deleteAll(notifications);
+                ArrayList<Notification> notifications = notificationRepository.findAllByStudent_Id(studentId);
+                notificationRepository.deleteAll(notifications);
 
-            ArrayList<Respond> responds = respondRepository.findAllByExecutor_Id(studentId);
-            respondRepository.deleteAll(responds);
+                ArrayList<Respond> responds = respondRepository.findAllByExecutor_Id(studentId);
+                respondRepository.deleteAll(responds);
 
-            studentRepository.deleteById(studentId);
-            return "success";
-
+                studentRepository.deleteById(studentId);
+                return "success";
+            } else {
+                return "У вас нет прав на данное действие";
+            }
         } catch (
                 Exception e) {
             return e.getMessage();
-        }
-
-    }
-
-    @DeleteMapping("/deleteNotification")
-    public String deleteNotification(Long notificationId) {
-        try {
-            notificationRepository.deleteById(notificationId);
-            return "success";
-        } catch (Exception e) {
-            return "e.getMessage()";
         }
     }
 }
